requirements.txt
# фреймворк
Django>=5
djangorestframework
drf-spectacular
django-debug-toolbar
pillow
# Пакеты для тестирования
pytest
pytest-django
pytest-cov
pytest-env

# Пакет для работы с env
python-dotenv
django-environ

# Для работы с БД Postgres
psycopg2-binary